requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
python-dotenv
googlemaps
Jinja2
google-generativeai